linkfinderbot/requirements.txt
asyncio
beautifulsoup4
lxml
playwright
python-telegram-bot==21.6
rapidfuzz